feat: use ABC for Action and Hook
