+++ incipyt/actions/__init__.py
@@ -1,5 +1,7 @@
 """TO-DO."""
 
+from incipyt.actions.abc import Action as _Action
+
 from incipyt.actions.git import Git
 from incipyt.actions.setuptools import Setuptools
 from incipyt.actions.venv import Venv
